d_a_npc_aj1: ctrl_WAITanm 98.64->100.0 + chk_parts_notMov s32->u8 (retail caller clrlwi.24) with u8 result var (s32 result var re-masks at return; direct returns restructure) -> 100.0; fresh .o postdates

+++ include/d/actor/d_a_npc_aj1.h
@@ -54,7 +54,7 @@ public:
     void eventOrder();
     void checkOrder();
     u8 chk_talk();
-    s32 chk_parts_notMov();
+    u8 chk_parts_notMov();
     void lookBack();
     s32 chkAttention();
     void setAttention(bool);
